Specimen Filter Builder

R6 class for building SQL queries to filter specimens with a fluent API

Methods

Public methods

SpecimenFilterBuilder$new()

SpecimenFilterBuilder$filter_collector()

SpecimenFilterBuilder$filter_number()

SpecimenFilterBuilder$filter_taxonomy()

SpecimenFilterBuilder$filter_by_ids()

SpecimenFilterBuilder$build()

SpecimenFilterBuilder$add_custom_condition()

SpecimenFilterBuilder$print_conditions()

SpecimenFilterBuilder$clone()

Method new()

Usage

SpecimenFilterBuilder$new(connection)

Method filter_collector()

Usage

SpecimenFilterBuilder$filter_collector(
  collector = NULL,
  id_colnam = NULL,
  interactive = FALSE
)

Method filter_number()

Usage

SpecimenFilterBuilder$filter_number(
  number = NULL,
  number_min = NULL,
  number_max = NULL
)

Method filter_taxonomy()

Usage

SpecimenFilterBuilder$filter_taxonomy(
  genus = NULL,
  species = NULL,
  family = NULL,
  idtax_n = NULL
)

Method filter_by_ids()

Usage

SpecimenFilterBuilder$filter_by_ids(specimen_ids)

Method build()

Usage

SpecimenFilterBuilder$build(operator = "AND")

Method add_custom_condition()

Usage

SpecimenFilterBuilder$add_custom_condition(condition, wrap_parentheses = TRUE)

Method print_conditions()

Usage

SpecimenFilterBuilder$print_conditions()

Method clone()

The objects of this class are cloneable with this method.

Usage

SpecimenFilterBuilder$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.